+++ src/main.rs
@@ -1,11 +1,12 @@
+mod analyzer;
 mod ast;
+mod bundler;
+mod env;
 mod errors;
 mod expr;
 mod lexer;
 mod parser;
-mod analyzer;
 mod resolver;
-mod env;
 #[cfg(test)]
 mod tests;