+++ changelog.md
@@ -1,3 +1,9 @@
 # Changelog
 
-## 1.0.0
+## 0.1.0
+
+- added bundler modules
+- updated ast
+- fixed resolver errors
+- fixed calls
+- fixed tests